Convert the rails_env variable to string
When using capistrano-torquebox in conjunction with capistrano-rails the
variable `rails_env` may be set to a symbol instead of a string.

If the user does not explicitly set `rails_env` it will be inferred from
the current Capistrano stage and in this case stages are represented by
symbols.

When we convert the deployment descriptor hash to yaml, YAML.dump turns
symbols like `:production` into the string `":production"` instead of
`"production"`

As a consequence, the -knob.yml deployment descriptor ends up with:

```
environment:
  RACK_ENV: :production
```

Which brakes the deploy because `":production"` is not the desired
environment, it's not even valid in most/all rails setups.